% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_roots_data.R
\name{check_format_roots_data}
\alias{check_format_roots_data}
\title{Prepare the root parameters for use un the Soil Water Balance model}
\usage{
check_format_roots_data(
  roots_data,
  layer_init_col = "depth_layer_init",
  layer_end_col = "depth_layer_end",
  percent_col = "percentRoots"
)
}
\arguments{
\item{roots_data}{data.frame, containing one-line-per-layer fine roots percentages
percentage as well as upper and lower depth bounds of the sampled layers (always positive numbers: low values being close to the surface, and high value being deep)}

\item{layer_init_col}{character, the name of the column containing upper (close to the surface) end depth for each layer}

\item{layer_end_col}{character, the name of the column containing lower (far from the surface) end depth for each layer}

\item{percent_col}{character, the name of the column containing observed fine roots percentage for each layer (percents, i.e. the sum of the column is 1)}
}
\value{
a list containing two elements, the first a data.frame containing modelled fine roots density for each 1cm sublayer, as well as information on
}
\description{
This function takes sampled root proportion among layers as an input dataset,
and uses a negative exponential function (/code{Rfd_{l} = /lambda
e^{-lambda*l}}) to estimate the distribution of fine roots among the soil, at
a 1cm resolution.
}
