% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_transpiration_parameters.R
\name{check_format_transpiration_parameters}
\alias{check_format_transpiration_parameters}
\title{Check and format understorey parameters for use in the Soil Water Balance
model}
\usage{
check_format_transpiration_parameters(...)
}
\arguments{
\item{...}{Either several scalars corresponding to the parameters meant to
replace default values in the model, or a single argument containing these
parameters in a grouped form (either list or single row data.frame). See
details for information on nomenclature and parameters.}
}
\value{
A named list containing the Understorey submodel parameters in this order: k, FractG, a, lambdaEU.
}
\description{
This function is initially meant for internal use.
}
\details{
The parameters are inputed either separately or groupedly.

 If inputed separately, each argument must have a name that corresponds to
 one of the Interception model's parameters. If the parameters are grouped
 into a list or a data.frame, the names of the columns/elements must also
 fulfill this requirement. Also note that if any parameter is missing, its
 corresponding default value will be used. Finally, if any argument/element
 has a name that does not correspond to any of the hereafter listed
 parameters, said argument will be ignored.
}
