% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_SWM.R
\name{display_SWM}
\alias{display_SWM}
\title{Display a single SWM simulation's output}
\usage{
display_SWM(
  model_output,
  what = "REW",
  type = "line",
  facet_by_time = "year",
  save_graph = F,
  path_save,
  filename,
  device,
  ...
)
}
\arguments{
\item{model_output}{data.frame and SWM object. The output of
Srun_original_SWM.}

\item{what}{The name of the variable to display, among "REW" and "Drainage"
(if relevant)}

\item{type}{The type of graph. It corresponds to a geom_ in ggplot's
namespace, e.g. "line" for geom_line, "point" for geom_point}

\item{facet_by_time}{character, or FALSE. If FALSE, the whole time series
will be displayed on a single graph (which can cause overplotting if it is
will be displayed on a single graph (which can cause overplotting if it is
long). Otherwise, please specify the name of the time column you wish to
use to facet (=break into different plots), e.g. "year", or "month" if you
have a single year. For multiple years to be displayed by month, you can
create a composite column (e.g. \code{year_month = paste(output$year,
output$month, sep = "_")}).}

\item{save_graph}{logical, whether the graph should be saved or not}

\item{path_save}{if save_graph = TRUE the location of the file}

\item{filename}{if save_graph = TRUE the name of the file}

\item{device}{if save_graph = TRUE, the device used (see \code{?ggplot2::ggsave()} for more details)}

\item{...}{additional graphical parameter for the corresponding geom.}
}
\value{

}
\description{
This function is meant to help the non-advanced R user to display the SWM
outputs in a nice way, using ggplot2. Graphical parameters can be further
tuned if specified in the optional arguments.
}
