% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_SWM.R
\name{display_SWM_stack}
\alias{display_SWM_stack}
\title{Display multiple SWM outputs}
\usage{
display_SWM_stack(
  model_stack,
  what = "REW",
  type = "line",
  facet_by_time = F,
  save_graph = F,
  path_save,
  filename,
  device,
  ...
)
}
\arguments{
\item{model_stack}{data.frame and SWM_stack. The output of stack_SWM, corresponding to daily aggregated outputs for several simulations.}

\item{what}{The name of the variable to display, among "REW" and "Drainage"
(if relevant)}

\item{type}{The type of graph. It corresponds to a geom_ in ggplot's
namespace, e.g. "line" for geom_line, "point" for geom_point}

\item{facet_by_time}{character, or FALSE. If FALSE, the whole time series
will be displayed on a single graph (which can cause overplotting if it is
will be displayed on a single graph (which can cause overplotting if it is
long). Otherwise, please specify the name of the time column you wish to
use to facet (=break into different plots), e.g. "year", or "month" if you
have a single year. For multiple years to be displayed by month, you can
create a composite column (e.g. \code{year_month = paste(output$year,
output$month, sep = "_")}).}

\item{save_graph}{logical, whether the graph should be saved or not}

\item{path_save}{if save_graph = TRUE the location of the file}

\item{filename}{if save_graph = TRUE the name of the file}

\item{device}{if save_graph = TRUE, the device used (see \code{?ggplot2::ggsave()} for more details)}

\item{...}{additional graphical parameter for the corresponding geom.}
}
\value{

}
\description{
THis function displays multiple SWM function outputs stacked using stack_SWM function.
}
