% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_original_SWM.R
\name{run_original_SWM}
\alias{run_original_SWM}
\title{Run the Original Code of SWM}
\usage{
run_original_SWM(
  rainfall_data = NULL,
  rainfall_col = NA,
  date_col = NA,
  date_format = NA,
  drainage = F,
  ...
)
}
\arguments{
\item{rainfall_data}{data.frame, but defaults to NA. Must contain daily
rainfall (if you have monthly rates, see convert_rainfall function), along
with a column containing dates (which format is specified with argument
date_format) or three column named "year","month" and "day" containing the
appropriate infos. If let to NA, a default dataset is automatically loaded
and the model is run.}

\item{rainfall_col}{character, but defaults to NA. If the column containing
rainfall in your dataset is named "pluie" as in Fabien's data, no need to
specify it.}

\item{date_col}{character, but defaults to NA. If this information is
contained in three columns names "year", "month", and "day" as in Fabien's
dataset, no need to specify it.}

\item{date_format}{single character, but defaults to NA. To be let to default
if date_col is also let to default. If your dataset contains a date column,
please 1- make sure the date format is strictly homogenous accross your
dataset ; and 2- specify it with three letters (lubridate's formalism): 'y'
for year, 'm' for month and 'd' for day, e.g.: 'ymd'.}

\item{drainage}{logical, defaults to FALSE. If TRUE, the SWM version "with
drainage" is run. I think the only difference is that deep drainage is also
outputed. To confirm with Fabien.}

\item{...}{}
}
\value{
A data.frame made from the input (or default) rainfall data,
containing the following columns: * date:The original or generated date
(according to how this information is included in the input dataset, see
above), parsed with lubridate according to its format (either specified
with date_format if you input complete dates as characters, or defaultly
working if your input dates are split into three columns, see above) *
year, month and day:The original or generated year, month and day for each
simulated date (according to how this information is included in the input
dataset, see above) * julian_day:The julian date corresponding to each
date, with origin set to 10-10-1960 as in Fabien's original codes. *
jour365The day indices computed from the 01-01 of the corresponding year *
rainfall, or whatever name inputed : the daily rainfall data given as an
input * REW: The computed Relative Extractable Water for trees accross the
soil profile (weighted by tree roots distribution accross the profile), for
each simulated day.
}
\description{
This function allow to run exactly the same code as Fabien does to output REW
values. Note that the C code called to run the simulations, and the
parameters of default simulations (used so far) are not exactly what is
decribed in the corresponding article. A document explaining the differences
between article and code will be edited, and available upon demand to
page.nino@gmail.com or geraldine.derroire@ecofog.gf
}
\details{
Here are a few notes to the attention of future users of this
package: * The original Soil Water Model is from Fabien Wagner (put
appropriate ref) and is exactly the code he used to provide REW and such
outputs up to now. This tool is meant to simplify the process of obtaining
such outputs, while letting open doors in order to allow some
parameter-tuning in case a few users would be tempted to explore it. *
Parameter-tuning must be done with caution. I would advise not to change
anything except rainfall data, and keep using this package on Paracou only,
because that is what Fabien advised us to do. * I noted several differences
between the algorithm, parameters and equations decribed in the article, and
those the C code called in this function actually does and uses. These differences will be documented as far as I can and will probably be available upon demand to Geraldine Derroire. I think this is
}
