% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_SWM_outputs.R
\name{stack_SWM}
\alias{stack_SWM}
\title{Stack SWM outputs}
\usage{
stack_SWM(..., SWM_names = NULL, time_correspondance = TRUE)
}
\arguments{
\item{...}{any objects outputed from run_original_SWM or stack_SWM}

\item{SWM_names}{Multiple character of same length as ... The names you want to attribute to each of the aforementioned object}

\item{time_correspondance}{Logical, do dates have to match perfectly ?}
}
\value{
A stack.
}
\description{
Stack SWM outputs
}
