% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_original_SWM.R
\name{SWMoutput_drainage}
\alias{SWMoutput_drainage}
\title{Title}
\usage{
SWMoutput_drainage(
  ThExp,
  ThetaPWP,
  ThetaFC,
  ThTr,
  N_layer,
  Pluie,
  EW_obs,
  jour,
  ProfMax,
  In,
  roots_couche,
  p,
  I0,
  LAI,
  ETP,
  LogL,
  N_day,
  ObsDay,
  L,
  ProfObs,
  Lobs,
  REW,
  percREW,
  EW_calc,
  Dr
)
}
\arguments{
\item{ThExp}{Single numeric. The exponential regression parameter used to
model fine roots density along the soil profile. Can be obtained with the
code used in check_format_roots_data}

\item{ThetaPWP}{Numeric. The median of ThetaPWP posterior distributions
obtained during model calibration, for each sampled depth.}

\item{ThetaFC}{Numeric. The median of ThetaFC posterior distributions
obtained during model calibration, for each sampled depth.}

\item{ThTr}{Numeric. The ratio of tree transpiration over PET. Referred to as
"rho" in the article. This ratio is decreased proportionally to REWc (a
measure of water load in a given layer) if it falls under a certain
threshold, that defines stressed hydraulic conditions}

\item{N_layer}{Sinle integer. The number of 1cm depth layers that are going
to be simulated.}

\item{Pluie}{Numeric. Rainfall data}

\item{EW_obs}{Numeric. Same length as real TDR sampled layers. Inputed as all
equal to 0.2. I think it is related to likelihood computation with the
markov chains algorithm, and thus deprecated or to be initialised with
dummy values?}

\item{jour}{Integer, =2. Used virtually nowhere in the code, not even
appearing in commented lines. Let's pretend it's better to keep it as is.}

\item{ProfMax}{Integer. The depth of the layers sampled with TDR. It may
correspond to centers rather than lower depth limits in the code that
expands soil parameters to 1cm-layers resolution.}

\item{In}{Numeric of the same length as Pluie. Interception values computed
for each day assuming one rainfall per day, and using the function
contained in m3.RData (but writted nowhere else in Fabien's folder)}

\item{roots_couche}{Negative integer. Basically equal to -1:-N_layer. Used in
the computation of understorey roots density.}

\item{p}{Numeric. A pretty mysterious vector of parameters, of which only a
part is used either in the model, or in the interception function. As for
the arguments of this function, nothing has been written anywhere to allow
external users to decode what it means.}

\item{I0}{Single numeric. 592.1, corresponds to mean daily incident radiation
to the canopy.}

\item{LAI}{Single numeric. Corresponds to PAI in the article, but it would
have been too easy and readable to name thing with at least a bit of
consistency.}

\item{ETP}{Numeric of the same length as Pluie. It originally contains only
3.970279 but it could take a daily timestep. Better not trying that.}

\item{LogL}{Zero. Perhaps log likelyhood, used in MCMC but useless here.}

\item{N_day}{Single integer. The number of days that are simulated. Must be
equal to the length of Pluie}

\item{ObsDay}{Same length as Pluie. Originally a vector of zeros containing a
single 1 at position 2800. Probably used to compute likelyhood.}

\item{L}{Single integer. The number of layers actually measured with TDR,
thus the number of values for soil parameters.}

\item{ProfObs}{Integer. Originally a vector of 1's of length 3. I genuinely
ignore what it's for. When it's equal to 0, does something with ThetaFC and
ThetaPWP values, but... Only Fabien could perhaps remember these kind of
very subtle details.}

\item{Lobs}{Single integer. The lines where it used to be implied have been
turned to comments. It seems that it's a deprecated argument replaced by L.
Perhaps it was used for calibrations, because it still is in the MCMCoutput
function written in C. Again, Fabien might be able to tell us more about
that than I'd ever be able to.}

\item{REW}{A vector of 0's of the same length as Pluie. I guess it's given as
an input to simplify dynamic memory allocation ? But I believe this could
be inclueded in the C code since this information is contained in other
arguments. Nothing to be modified here except the length of this arg.}

\item{percREW}{Zero... Same remark here. Why isn't its declaration and
instance into the C code?? Also not to be modified.}

\item{EW_calc}{Idem.}

\item{Dr}{Idem}
}
\value{

}
\description{
Title
}
